🚑 fix: Update AuditSidebar component to use new encryptPDF function
- Updated AuditSidebar component to utilize the updated `encryptPDF` function from the audits service, which now includes additional parameters such as `auditId`, `auditName`, `setIsOpenModal`, and `setIsGeneratingPDF`. This change ensures proper encryption of PDF files when handling audit-related functionalities.
