+++ frontend/src/components/navbar/AuditSidebar.tsx
@@ -6,7 +6,7 @@ import { Link, useParams } from 'react-router-dom';
 import { toast } from 'sonner';
 
 import { EncryptionModal } from '@/routes/audits/edit/general/EncryptionModal';
-import { getAuditById } from '@/services/audits';
+import { encryptPDF, getAuditById } from '@/services/audits';
 
 import DefaultRadioGroup from '../button/DefaultRadioGroup';
 import DropdownButton, { ListItem } from '../button/DropdownButton';
@@ -174,44 +174,16 @@ const AuditSidebar = ({
   });
 
   const handleSubmitEncrypt = async (password: string) => {
-    const bodyParam = {
-      password,
-    };
-    setIsGeneratingPDF(true);
-
     try {
-      const response = await fetch(
-        `${import.meta.env.VITE_API_URL}/api/audits/${auditId}/generate/pdf`,
-        {
-          credentials: 'include',
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify(bodyParam),
-        },
+      await encryptPDF(
+        password,
+        auditId ?? '',
+        auditName,
+        setIsOpenModal,
+        setIsGeneratingPDF,
       );
-
-      if (!response.ok) {
-        throw new Error('Error generating PDF');
-      }
-
-      const blob = await response.blob();
-      const url = window.URL.createObjectURL(blob);
-      const link = document.createElement('a');
-      link.href = url;
-      link.download = `${auditName}.pdf`;
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-      window.URL.revokeObjectURL(url);
-
-      setIsOpenModal(false);
-    } catch (error) {
-      console.error('Error generating PDF:', error);
+    } catch (_) {
       toast.error(t('err.errorGeneratingPdf'));
-    } finally {
-      setIsGeneratingPDF(false);
     }
   };
